Extract SearchIndexUpdates into standalone component in vector section
- Move impact bar logic out of WriteTripleForm into new SearchIndexUpdates component
- Place SearchIndexUpdates below WriteTripleForm inside VectorPipelineCard
- Add onWriteComplete callback to WriteTripleForm to pass mzLowerBound + wallClock
- Highlight embedding_text block with yellow glow when embedding re-indexes
- Install typescript devDep so tsc build step works
