+++ web/package.json
@@ -42,7 +42,7 @@
     "jsdom": "^23.2.0",
     "postcss": "^8.4.33",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.3.3",
+    "typescript": "^5.9.3",
     "vite": "^5.0.11",
     "vitest": "^1.2.0"
   }
