+++ web/src/components/WriteTripleForm.tsx
@@ -1,23 +1,6 @@
 import { useState, useEffect, useMemo } from "react";
 import { Edit3 } from "lucide-react";
-import { queryStatsApi, searchApi } from "../api/client";
-import { usePropagation } from "../contexts/PropagationContext";
-
-const VIRTUAL_SIZE = 65536; // large space so marks are proportionally sparse
-
-function hashDocId(id: string): number {
-  let h = 0;
-  for (let i = 0; i < id.length; i++) {
-    h = Math.imul(31, h) + id.charCodeAt(i) >>> 0;
-  }
-  return h % VIRTUAL_SIZE;
-}
-
-const OP_COLORS: Record<string, string> = {
-  INSERT: '#10b981',
-  UPDATE: '#a855f7',
-  DELETE: '#ef4444',
-};
+import { queryStatsApi } from "../api/client";
 
 const predicatesBySubjectType: Record<string, string[]> = {
   order: ['order_status', 'order_number', 'delivery_window_start', 'delivery_window_end'],
@@ -46,42 +29,16 @@ const placeholdersByPredicate: Record<string, string> = {
 interface WriteTripleFormProps {
   initialSubject?: string;
   onWritten?: () => void;
+  onWriteComplete?: (mzLowerBound: number, wallClock: number) => void;
 }
 
-export const WriteTripleForm = ({ initialSubject = "", onWritten }: WriteTripleFormProps) => {
+export const WriteTripleForm = ({ initialSubject = "", onWritten, onWriteComplete }: WriteTripleFormProps) => {
   const [subject, setSubject]     = useState(initialSubject);
   const [predicate, setPredicate] = useState("quantity");
   const [value, setValue]         = useState("");
   const [status, setStatus]       = useState<string | null>(null);
-  const [impact, setImpact]       = useState<{ impacted: number; total: number; pct: number } | null>(null);
-  const [watchingImpact, setWatchingImpact] = useState(false);
-  const [writeTimestamp, setWriteTimestamp] = useState<number | null>(null);
-
-  const { events } = usePropagation();
-
-  // Marks: one per unique doc_id changed since last write, at a proportional position
-  const marks = useMemo(() => {
-    if (!writeTimestamp) return null;
-    const relevant = events.filter((e) => e.timestamp >= writeTimestamp);
-    if (relevant.length === 0) return null;
-    const seen = new Map<string, string>(); // doc_id → color
-    for (const e of relevant) {
-      seen.set(e.doc_id, OP_COLORS[e.operation] ?? '#6b7280');
-    }
-    return Array.from(seen.entries()).map(([doc_id, color]) => ({
-      doc_id,
-      color,
-      pct: (hashDocId(doc_id) / VIRTUAL_SIZE) * 100,
-    }));
-  }, [events, writeTimestamp]);
 
   useEffect(() => { setSubject(initialSubject); }, [initialSubject]);
-
-  // Clear impact + buckets whenever the user starts a new search
-  useEffect(() => {
-    setImpact(null);
-    setWriteTimestamp(null);
-  }, [subject, predicate, value]);
 
   const availablePredicates = useMemo(() => {
     let base = predicatesBySubjectType.orderline;
@@ -101,25 +58,6 @@ export const WriteTripleForm = ({ initialSubject = "", onWritten }: WriteTripleF
     }
   }, [availablePredicates, predicate]);
 
-  const watchImpact = async (lowerBound: number) => {
-    setWatchingImpact(true);
-    setImpact(null);
-    let lastImpacted = -1;
-    for (let i = 0; i < 8; i++) {
-      await new Promise(r => setTimeout(r, 1000));
-      try {
-        const res = await searchApi.indexImpact(lowerBound);
-        const data = res.data;
-        if (data.impacted > 0 && data.impacted === lastImpacted) break;
-        lastImpacted = data.impacted;
-        setImpact(data);
-      } catch {
-        break;
-      }
-    }
-    setWatchingImpact(false);
-  };
-
   const handleWrite = async () => {
     if (!subject || !predicate || !value) return;
     if (subject.length > 255) { flash("Error: Subject too long"); return; }
@@ -128,15 +66,12 @@ export const WriteTripleForm = ({ initialSubject = "", onWritten }: WriteTripleF
     if (!subject.includes(':') && !subject.includes('_')) {
       flash("Error: Subject should be 'type:id' or 'type_id'"); return;
     }
-    setImpact(null);
-    setWriteTimestamp(null);
     try {
       const res = await queryStatsApi.writeTriple({ subject_id: subject, predicate, object_value: value });
-      setWriteTimestamp(Date.now() / 1000); // seconds to match PropagationEvent.timestamp
       flash(`Written at ${new Date().toLocaleTimeString()}`);
       onWritten?.();
       if (res.data.mz_timestamp_lower_bound != null) {
-        watchImpact(res.data.mz_timestamp_lower_bound);
+        onWriteComplete?.(res.data.mz_timestamp_lower_bound, Date.now() / 1000);
       }
     } catch {
       flash("Write failed");
@@ -202,44 +137,6 @@ export const WriteTripleForm = ({ initialSubject = "", onWritten }: WriteTripleF
           </span>
         )}
       </div>
-      <div className="mt-3">
-          {/* Proportional marker bar — always visible, markers fill in after write */}
-          <div className="relative h-4 rounded overflow-hidden bg-gray-200">
-            {marks?.map(({ doc_id, color, pct }) => (
-              <div
-                key={doc_id}
-                className="absolute top-0 bottom-0"
-                style={{ left: `${pct}%`, width: '2px', backgroundColor: color }}
-                title={doc_id}
-              />
-            ))}
-          </div>
-          {/* Count row — only shown after a write */}
-          {(watchingImpact || impact) && (
-          <div className="mt-1 text-xs text-gray-500 flex items-center gap-2">
-            {watchingImpact && !impact && (
-              <span className="text-gray-400 animate-pulse">Watching pipeline…</span>
-            )}
-            {impact && (
-              <>
-                <span className="font-mono font-semibold text-purple-700">{impact.impacted}</span>
-                <span className="text-gray-400">/</span>
-                <span className="font-mono text-gray-600">{impact.total}</span>
-                <span>docs re-indexed</span>
-                <span className={`font-semibold ${impact.pct > 10 ? 'text-orange-600' : 'text-purple-600'}`}>
-                  ({impact.pct}%)
-                </span>
-                {watchingImpact && <span className="text-gray-400 animate-pulse">…</span>}
-              </>
-            )}
-            <div className="ml-auto flex items-center gap-3 text-gray-400">
-              <span className="flex items-center gap-1"><span className="inline-block w-2 h-2 rounded-sm" style={{ backgroundColor: OP_COLORS.INSERT }} />insert</span>
-              <span className="flex items-center gap-1"><span className="inline-block w-2 h-2 rounded-sm" style={{ backgroundColor: OP_COLORS.UPDATE }} />update</span>
-              <span className="flex items-center gap-1"><span className="inline-block w-2 h-2 rounded-sm" style={{ backgroundColor: OP_COLORS.DELETE }} />delete</span>
-            </div>
-          </div>
-          )}
-        </div>
     </div>
   );
 };
